feat(tests): reorganize test structure and enhance configurations
- Updated Playwright configuration to reflect new test directory structure.
- Expanded Steiger configuration to ignore additional test artifacts.
- Refactored Vitest setup to point to the new shared tests directory.
- Introduced new test files for integration and end-to-end testing.
- Added mock handlers for API responses and improved test coverage for UI components.
- Created stories for various UI components to facilitate visual testing.
