+++ playwright.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig, devices } from "@playwright/test";
 
 export default defineConfig({
-  testDir: "./e2e",
+  testDir: "./src/app/tests/e2e",
   fullyParallel: true,
   retries: process.env.CI ? 2 : 0,
   reporter: "html",
