+++ src/app/tests/index.ts
@@ -1,0 +1,1 @@
+export { handlers } from "../tests/mocks/handlers";
