+++ src/shared/tests/index.ts
@@ -1,0 +1,1 @@
+export { createTestProviders, renderWithProviders } from "../tests/render";
