+++ src/vite-env.d.ts
@@ -1,7 +1,12 @@
 /// <reference types="vite/client" />
 
 interface ImportMetaEnv {
+  readonly BASE_URL: string;
   readonly VITE_API_BASE_URL: string;
   readonly VITE_APP_BASE_URL: string;
   readonly VITE_ENABLE_MSW: string;
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv;
 }
